+++ x/wasm/internal/keeper/keeper_test.go
@@ -8,8 +8,6 @@ import (
 	"testing"
 	"time"
 
-	"github.com/CosmWasm/go-cosmwasm"
-	wasmTypes "github.com/CosmWasm/go-cosmwasm/types"
 	"github.com/CosmWasm/wasmd/x/wasm/internal/types"
 	stypes "github.com/cosmos/cosmos-sdk/store/types"
 	sdk "github.com/cosmos/cosmos-sdk/types"
@@ -436,33 +434,16 @@ func TestInstantiateWithNonExistingCodeID(t *testing.T) {
 func TestInstantiateWithCallbackToContract(t *testing.T) {
 	ctx, keepers := CreateTestInput(t, false, SupportedFeatures, nil, nil)
 	var (
-		excuteCalled bool
-		err          error
+		executeCalled bool
+		err           error
 	)
-	wasmerMock := &MockWasmer{
-		CreateFn: func(code cosmwasm.WasmCode) (cosmwasm.CodeID, error) {
-			anyCodeID := bytes.Repeat([]byte{0x1}, 32)
-			return anyCodeID, nil
-		},
-		InstantiateFn: func(code cosmwasm.CodeID, env wasmTypes.Env, info wasmTypes.MessageInfo, initMsg []byte, store cosmwasm.KVStore, goapi cosmwasm.GoAPI, querier cosmwasm.Querier, gasMeter cosmwasm.GasMeter, gasLimit uint64) (*wasmTypes.InitResponse, uint64, error) {
-			return &wasmTypes.InitResponse{
-				Messages: []wasmTypes.CosmosMsg{
-					{Wasm: &wasmTypes.WasmMsg{Execute: &wasmTypes.ExecuteMsg{ContractAddr: env.Contract.Address, Msg: []byte(`{}`)}}},
-				},
-			}, 1, nil
-		},
-		ExecuteFn: func(code cosmwasm.CodeID, env wasmTypes.Env, info wasmTypes.MessageInfo, executeMsg []byte, store cosmwasm.KVStore, goapi cosmwasm.GoAPI, querier cosmwasm.Querier, gasMeter cosmwasm.GasMeter, gasLimit uint64) (*wasmTypes.HandleResponse, uint64, error) {
-			excuteCalled = true
-			return &wasmTypes.HandleResponse{}, 1, nil
-		},
-	}
+	wasmerMock := selfCallingInstMockWasmer(&executeCalled)
 
-	keepers.WasmKeeper.wasmer = wasmerMock
 	keepers.WasmKeeper.wasmer = wasmerMock
 	example := StoreHackatomExampleContract(t, ctx, keepers)
 	_, err = keepers.WasmKeeper.Instantiate(ctx, example.CodeID, example.CreatorAddr, nil, nil, "test", nil)
 	require.NoError(t, err)
-	assert.True(t, excuteCalled)
+	assert.True(t, executeCalled)
 }
 
 func TestExecute(t *testing.T) {
